added auth to post, as well as some logic to keep you from ever getting to the postArticle page without a valid ID, but you can get there. So must have valid idToken to post, which is variable
