+++ next.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  env: {
+    post: "AIzaSyDcnw_qBhLHxj3mngZQBXMrlQCU7_I4l8g",
+    cookie: "laj_alj4o5892-mfhgr_efijogrgsrg",
+    secret: "UOb6bhxsFtdyDjgfcZ128FyS14d2",
+    url: "https://evron-dev-blog-default-rtdb.firebaseio.com/newPosts.json",
+  },
+};
