+++ pages/post/index.js
@@ -1,223 +1,69 @@
+import { useEffect, useState } from "react";
 import Layout from "../../Components/Layout/Layout";
-import React, { useState, useEffect, useContext } from "react";
-import { ModeContext } from "../../Context/context";
+import axios from "axios";
 import { useRouter } from "next/router";
-import styles from "./post.module.css";
 
-const Post = (props) => {
-  const mainContext = useContext(ModeContext);
+const Login = () => {
   const router = useRouter();
-  const [submitObj, setSubmitObj] = useState({
-    author: "",
-    title: "",
-    quip: "",
-    tags: [],
-    mainImg: "",
-  });
-  const [numTags, setNumTags] = useState(0);
-  const [final, setFinal] = useState();
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
 
   useEffect(() => {
-    if (final) {
-      mainContext.submitDataSet(final);
-      router.push({
-        pathname: "/post/check",
-      });
-    }
-    return;
-  }, [final]);
+    if (localStorage.getItem(process.env.cookie) === process.env.secret)
+      return router.push("/post/postArticle");
+  }, []);
+
+  const inputHandler = (e) => {
+    e.target.id === "email"
+      ? setEmail(e.target.value)
+      : setPassword(e.target.value);
+  };
 
   const submitHandler = (e) => {
     e.preventDefault();
-    const almostFinal = [];
-    let group = 0;
-    for (let i = 12; i < e.target.length - 1; i++) {
-      if (i % 2 === 0) {
-        almostFinal.push({ type: e.target[i].value });
-      }
-      if (i % 2 === 1) {
-        almostFinal[group].content = e.target[i].value;
-        group++;
-      }
-    }
-    let final = {
-      ...submitObj,
-      body: almostFinal,
-    };
-    setFinal(final);
+
+    axios
+      .post(
+        `https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=${process.env.post}`,
+        { email, password, returnSecureToken: true }
+      )
+      .then((res) => {
+        console.log(res);
+        localStorage.setItem(process.env.cookie, res.data.localId);
+        localStorage.setItem("cookieId", res.data.idToken);
+        localStorage.setItem("cookieRefresh", res.data.refreshToken);
+        router.push("/post/postArticle");
+      })
+      .catch((error) => console.log(error));
   };
-
-  //switch to switch statement
-  const mainInputCHangehandler = (e) => {
-    if (e.target.id === "title") {
-      setSubmitObj((prevState) => {
-        let state = { ...prevState, title: e.target.value };
-        return state;
-      });
-    }
-    if (e.target.id === "author") {
-      setSubmitObj((prevState) => {
-        let state = { ...prevState, author: e.target.value };
-        return state;
-      });
-    }
-    if (e.target.id === "quip") {
-      setSubmitObj((prevState) => {
-        let state = { ...prevState, quip: e.target.value };
-        return state;
-      });
-    }
-    if (e.target.id === "mainImg") {
-      setSubmitObj((prevState) => {
-        let state = { ...prevState, mainImg: e.target.value };
-        return state;
-      });
-    }
-  };
-  const numTagsHandler = (e) => {
-    setNumTags(e.target.value);
-  };
-
-  const checkboxHandler = (e) => {
-    if (e.target.checked) {
-      setSubmitObj((prevState) => {
-        prevState.tags = [...prevState.tags, e.target.value];
-        return prevState;
-      });
-    }
-    if (!e.target.checked) {
-      setSubmitObj((prevState) => {
-        prevState.tags = prevState.tags.filter((item) => {
-          return item !== e.target.value;
-        });
-        return prevState;
-      });
-    }
-  };
-
-  // i guess the virtual dom sees only a few change, so it doesnt re-render items that stayw
-  let displayBodyInputs = [];
-
-  if (numTags) {
-    for (let i = 0; i < numTags; i++) {
-      displayBodyInputs.push(
-        <div style={{ width: "700px" }}>
-          <select>
-            <option value="p">p</option>
-            <option value="img">img</option>
-            <option value="code">code</option>
-            <option value="sub">sub-title</option>
-          </select>
-          <textarea
-            style={{ minWidth: "700px", maxWidth: "700px" }}
-            id={`bi${i}`}
-            type="text"
-          />
-        </div>
-      );
-    }
-  }
-
   return (
-    <Layout title="Post Private">
-      <div className={styles.postContainer}>
-        <form onSubmit={submitHandler} className={styles.form}>
-          <label htmlFor="title">Author </label>
-          <select
-            id="author"
-            onChange={mainInputCHangehandler}
-            value={submitObj.author}
-          >
-            <option value=""></option>
-            <option value="jon">Jon</option>
-            <option value="mikey">Mikey</option>
-          </select>
-          <label htmlFor="title">Title </label>
-          <input
-            id="title"
-            type="text"
-            placeholder="title"
-            value={submitObj.title}
-            onChange={mainInputCHangehandler}
-          ></input>
-          <label htmlFor="title">Quip </label>
-          <input
-            id="quip"
-            type="text"
-            placeholder="quip"
-            value={submitObj.quip}
-            onChange={mainInputCHangehandler}
-          ></input>
-          <label htmlFor="mainImg">Main Img </label>
-          <input
-            id="mainImg"
-            type="text"
-            placeholder="mainImg"
-            value={submitObj.mainImg}
-            onChange={mainInputCHangehandler}
-          ></input>
-          {/* Hidden array used to even out loop */}
-          <input id="hidden" type="hidden" />
-          <div className={styles.techs}>
-            <label htmlFor="react">React</label>
+    <Layout>
+      <section>
+        <h2>Login</h2>
+        <form className="form" onSubmit={submitHandler}>
+          <div className="email">
+            <label htmlFor="email">Email</label>
             <input
-              type="checkbox"
-              value="React"
-              id="react"
-              onChange={checkboxHandler}
-            />
-            <label htmlFor="next">Next</label>
-            <input
-              type="checkbox"
-              value="Next"
-              id="next"
-              onChange={checkboxHandler}
-            />
-            <label htmlFor="node">Node</label>
-            <input
-              type="checkbox"
-              value="Node"
-              id="node"
-              onChange={checkboxHandler}
-            />
-            <label htmlFor="js">JS</label>
-            <input
-              type="checkbox"
-              value="JS"
-              id="js"
-              onChange={checkboxHandler}
-            />
-
-            <label htmlFor="html5">Html5</label>
-            <input
-              type="checkbox"
-              value="Html5"
-              id="html5"
-              onChange={checkboxHandler}
-            />
-            <label htmlFor="css3">Css3</label>
-            <input
-              type="checkbox"
-              value="Css3"
-              id="css3"
-              onChange={checkboxHandler}
+              id="email"
+              type="email"
+              placeholder="email"
+              onChange={inputHandler}
             />
           </div>
-          <div>
-            <label htmlFor="numTags">How many Tags?</label>
+          <div className="password">
+            <label htmlFor="password">Password</label>
             <input
-              id="numTags"
-              type="number"
-              min="0"
-              onChange={numTagsHandler}
-            ></input>
+              id="password"
+              type="password"
+              placeholder="password"
+              onChange={inputHandler}
+            />
           </div>
-          <div>{displayBodyInputs}</div>
-          <button>Submit Form</button>
+          <button>Submit</button>
         </form>
-      </div>
+      </section>
     </Layout>
   );
 };
 
-export default Post;
+export default Login;
