+++ utils/postData.js
@@ -1,11 +1,15 @@
 const axios = require("axios");
 
 const post = async (item) => {
-  let final = item;
-  await axios.post(
-    "https://evron-dev-blog-default-rtdb.firebaseio.com/newPosts.json",
-    final
-  );
+  try {
+    let final = item;
+    await axios.post(
+      `${process.env.url}?auth=${localStorage.getItem("cookieId")}`,
+      final
+    );
+  } catch (e) {
+    console.log(e);
+  }
 };
 
 export default post;
